Move Assassin -> Resurrect as Police to PolicePlugin
This is so that it is not active during games without permadeath.
