+++ AU2/plugins/CorePlugin.py
@@ -117,13 +117,6 @@ class CorePlugin(AbstractPlugin):
                 ((lambda: ASSASSINS_DATABASE.get_identifiers()),)
             ),
             Export(
-                "core_assassin_to_police",
-                "Assassin -> Resurrect as Police",
-                self.ask_core_plugin_assassin_to_police,
-                self.answer_core_plugin_assassin_to_police,
-                (self.gather_dead_non_police,)
-            ),
-            Export(
                 "core_event_create_event",
                 "Event -> Create",
                 self.ask_core_plugin_create_event,
@@ -407,29 +400,6 @@ class CorePlugin(AbstractPlugin):
         for p in PLUGINS:
             return_components += p.on_assassin_update(assassin, html_response_args)
         return return_components
-
-    def gather_dead_non_police(self) -> List[str]:
-        # use targeting plugin as proxy for whether we have perma-death
-        targeting_enabled = any(p.identifier == "TargetingPlugin" for p in PLUGINS)
-        death_manager = DeathManager(perma_death=targeting_enabled)
-        for e in EVENTS_DATABASE.events.values():
-            death_manager.add_event(e)
-        return ASSASSINS_DATABASE.get_identifiers(include=(lambda a: death_manager.is_dead(a) and not a.is_police))
-
-    def ask_core_plugin_assassin_to_police(self, ident: str):
-        components = [HiddenTextbox(identifier=self.HTML_SECRET_ID, default=ident),
-                      NamedSmallTextbox(identifier=self.html_ids["Pseudonym"], title="New initial pseudonym")]
-        return components
-
-    def answer_core_plugin_assassin_to_police(self, html_response_args: Dict):
-        ident = html_response_args[self.HTML_SECRET_ID]
-        assassin = ASSASSINS_DATABASE.get(ident)
-        new_pseudonym = html_response_args[self.html_ids["Pseudonym"]]
-        new_assassin = assassin.clone(hidden=False, is_police=True, pseudonyms=[new_pseudonym])
-        ASSASSINS_DATABASE.add(new_assassin)
-        # hide the old assassin
-        assassin.hidden = True
-        return [Label(f"[CORE] Resurrected {ident} as {new_assassin.identifier}.")]
 
     def ask_core_plugin_create_event(self):
         components = []
